requirements.txt
openpyxl==3.1.2
Pillow==9.5.0
Send2Trash==1.8.0
